# Conflicts:
#	UWP/Renderer/XamlCardRenderer.vcxproj
#	UWP/Renderer/XamlCardRenderer.vcxproj.filters
#	UWP/Renderer/idl/AdaptiveCards.XamlCardRenderer.idl
#	UWP/Renderer/lib/AdaptiveContainer.cpp
#	UWP/Renderer/lib/ImageLoadTracker.cpp
#	UWP/Renderer/lib/ImageLoadTracker.h
#	UWP/Renderer/lib/XamlBuilder.cpp
#	UWP/Renderer/lib/XamlBuilder.h
#	UWP/Renderer/lib/XamlCardRendererComponent.cpp
#	UWP/Renderer/lib/XamlCardRendererComponent.h
#	UWP/Renderer/lib/XamlHelpers.h
#	UWP/Renderer/lib/XamlStyleKeyGenerators.h
#	UWP/Visualizer/MainPage.xaml.cs

+++ UWP/Renderer/lib/Util.cpp
@@ -21,7 +21,7 @@ HRESULT UTF8ToHString(const string& in, HSTRING* out)
     }
     wstring_convert<codecvt_utf8_utf16<wchar_t>> converter;
     wstring wide = converter.from_bytes(in);
-    return WindowsCreateString(wide.c_str(), wide.length(), out);
+    return WindowsCreateString(wide.c_str(), static_cast<UINT32>(wide.length()), out);
 }
 
 HRESULT HStringToUTF8(const HSTRING& in, string& out)
